requirements.txt
asgiref==3.10.0
attrs==25.4.0
black==25.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
Django==5.2.7
django-cors-headers==4.9.0
django-elasticsearch-dsl==9.0
django-import-export==4.1.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
elastic-transport==9.2.0
elasticsearch==9.1.1
flake8==7.3.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
isort==7.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytokens==0.2.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rest-framework-simplejwt==0.0.2
rpds-py==0.28.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0

